+++ src/optionValidator.ts
@@ -1,7 +1,12 @@
 import {isPlainObject} from 'lodash'
 import {Options} from '.'
 
-export function validateOptions({formatTypes, maxItems}: Partial<Options>): void {
+export function validateOptions({declarationStyle, formatTypes, maxItems}: Partial<Options>): void {
+  if (declarationStyle !== undefined && declarationStyle !== 'interface' && declarationStyle !== 'type') {
+    throw RangeError(
+      `Expected options.declarationStyle to be "interface" or "type", but was given ${declarationStyle}.`,
+    )
+  }
   if (
     formatTypes !== undefined &&
     !(isPlainObject(formatTypes) && Object.values(formatTypes).every(_ => typeof _ === 'string'))